Title: 세 수의 합과 평균 구하기, Time: 132ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240205/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -8,6 +8,6 @@ public class Main {
         int c = sc.nextInt();
 
         System.out.println(a+b+c);
-        System.out.println((double)(a+b+c)/3);
+        System.out.println((a+b+c)/3);
     }
 }
